r9563/sofa : FIX: many warnings on windows, so that relevant ones can be taken into account Also added a few @todo where I believe some code is incorrect (to be checked later)
Former-commit-id: a68041640ae9044d00f709e5da70607a3dc8cd81

+++ applications/plugins/SofaHAPI/SofaHAPIHapticsDevice.h
@@ -76,8 +76,8 @@ public:
     Data<double> toolTransitionSpringStiffness;
     Data<std::string> driverName;
     Data<bool> drawDevice;
-    Data<double> drawHandleSize;
-    Data<double> drawForceScale;
+    Data<float> drawHandleSize;
+    Data<float> drawForceScale;
 
     //SofaHAPIHapticsDeviceData data;
     ForceFeedbackTransform data;
